vim: First try on omni completion for account names
That is a completion which is aware of what it should complete.
Currently only account names are supported.
When you insert an account name like this:
  Asse<C-X><C-O>
You will get a list of top-level accounts that start like this.
Go ahead and try something like:
  As:Ban:Che<C-X><C-O>
When you have an account like this, 'Assets:Bank:Checking'
should show up.