Add functions to print warning/error messages.
Add functions to print warnings and error messages below the status
line.

The first thing these functions do is calling `redraw`. This is
necessary in some circumstances to avoid messages being chained together
and causing Vim to bother the user with ‘Press ENTER or type command to
continue’. Similarly, the `echo` command is needed to move the cursor
to the first column: there are situations when the message would be
appended to existing text if the cursor is not moved to the first
column (this happens, for example, if an echo command is given after
`input()`).