Send register reports to the quickfix list.
Add a `:Register` command to send a register report to the quickfix
window. This is very handy because the entries in the report and the
postings in the source files are synchronized (you may jump from an
entry in the quickfix window to the corresponding location in the source
code).