Display account balance below the status line.
Add a `:Balance` command to display an account’s cleared and pending
balance below the status line. The command takes an account name as an
optional argument: if no account is specified, the account in the
current line is used.

The balance line can be fully customized: highlight groups are defined
for the colors and global variables for the text.