Use the last line of output in #show_balance().
We first split on line breaks and only use the last line of ledger’s
output in order to get rid of warning messages etc. produced by ledger
while parsing the file.

We also check for errors: for example, an account with multiple
commodities will produce a format error in Ledger if `-X` is not used.