Use `-f` in #reconcile() and #show_balance().
Add `-f` to those functions so that they are applied to the correct
file.

Add a global variable to specify the main ledger file (similar in
purpose to the LEDGER_FILE environment variable of Ledger). By default,
this is set to expand to the current buffer.

Add an optional file path parameter to those functions. If a file path
is not provided explicitly then the value of g:ledger_main is used.