Use `command! -buffer` to define the commands.
Use `:command!` to overwrite existing definitions (and remove the if
condition) while using `-buffer` to restrict the commands to the ledger
buffer (see :help :command-buffer). This change makes
`is_ledger_buffer()` useless.

Add a `-f` argument to the `:Ledger` and `:Register` commands (using the
value of `g:ledger_main`), so that the user does not need to specify it
on the command line.