Use g:ledger_main when autocompleting commands.
The account name auto-completion does not work when editing a file that
does not balance on its own (same situation with 2015.ldg and main.ldg). Use expand(g:ledger_main) instead of expand('%') to fix that.

Apart from that, `s:autocomplete_account_or_payee` does work correctly
when Ledger spits out warning messages or errors. This will be addressed
in a future commit.