Add g:ledger_extra_options.
The purpose of this variable is to allow the user to specify default
options to pass to ledger when generating reports. Previously, such
options would have to be included in `g:ledger_bin`, which would make it
difficult to call ledger without including them (which in some cases is
desirable, e.g., in `ledger#entry()` or when getting the list of payees
or accounts).

So, from now on `g:ledger_bin` is assumed to be the path to the ledger
executable, without further options.

The two global variables are now documented.