Simplify the parsing of Ledger commands.
Ledger uses `%` in format strings. The current way of processing Ledger
commands makes it impossible for the user to specify format strings
correctly, because a `%` preceded by a space is expanded to the current
buffer's name. Try, for example:

:Ledger bal -F ' %(amount)'

Note that escaping (as suggested in the comment of s:ledger_cmd()) does
not work either. The following:

:Ledger bal -F ' \%(amount)'

becomes

ledger bal -F ' \%(amount)'

that is, the format is passed as is and ledger sees the backslash. There
may be similar situations with other symbols (e.g, when using `<`).

This commit fixes these problems. Instead of trying to be clever and try
to expand file paths (e.g., `%`) automatically, pass the file to be
processed as an explicit argument and do not expand the user’s
arguments.