Refactor ledger#entry()
The entry front-end was buggy because of how it escaped the query. It
was fixed by refactoring to use `systemlist(s:ledger_cmd())` instead of
`read !` with `shellescape()`.

If the call to `ledger entry` fails, it echoes the failure message, and
leaves the input query to be amended by the user.