allow completion from declared accounts
When using ledger with the `--strict` option, you need to declare all
your used accounts. These declarations can additionally be used to
complete account names.

ledger#declared_accounts() works a lot like ledger#transactions(), on
which it is based. s:collect_completion_data() can then take the list of
declared accounts as an initial value to which it then adds the accounts
used in transactions.

Signed-off-by: Roland Hieber <rohieb@rohieb.name>