Fix makeprg to work around ledger cli bug
Using `ledger -f % source %` shouldn't be necessary and the value of the
second % is actually ignored, but it doesn't work without it. That bit
should be removed when the upstream bug is fixed:

https://github.com/ledger/ledger/issues/1682

The rest of this just clean up the detection of ledger vs. hledger now
that we support both so this only gets set when it will be expected to
work.