fix(foldtext): handle negative balance assertions (#165)
Further expand the regular expressions to match optional balance
assertions to catch negative balance assertions.

Amounts can be signed with a negative symbol, or a positive one;
the currency symbol and a sign symbol can be in either order:
https://hledger.org/1.41/hledger.html#amounts

Add a test ledger under test/ to demonstrate some of the
combinations.

Signed-off-by: Jonathan Dowland <jon@dow.land>