fix: Disable formatprg by default to avoid potential data loss
Closes #168

For those that want to live dangerously and enable this by default, one
can set a truthy value for `g:ledger_dangerous_formatprg` to re-enable
the default ledger print command as a formatter. Just be aware that
should only be run on transactions, not other statement types or comment
blocks.